+++ R/mo_property.R
@@ -224,6 +224,8 @@ mo_shortname <- function(x, language = get_AMR_locale(), keep_synonyms = getOpti
   # unknown species etc.
   shortnames[shortnames %like% "unknown"] <- paste0("(", trimws2(gsub("[^a-zA-Z -]", "", shortnames[shortnames %like% "unknown"], perl = TRUE)), ")")
 
+  shortnames[mo_rank(x.mo) %in% c("kingdom", "phylum", "class", "order", "family")] <- mo_name(x.mo, language = NULL, keep_synonyms = keep_synonyms)
+  
   shortnames[is.na(x.mo)] <- NA_character_
   load_mo_uncertainties(metadata)
   translate_into_language(shortnames, language = language, only_unknown = FALSE, only_affect_mo_names = TRUE)